First commit
This is by no means a working piece of software yet. The main structure
is there, but many bolts and pieces are still missing. Some important
points though:

 * The main PSO loop is sam-agnostic, in a way, just like in the
original code. A "SAM runner" object is created depending on the SAM
that one wants to PSO, which internally implements the logic to queue
the SAM, wait for results, read the relevant data back, etc. Different
SAM runner classes will thus implement things differently depending on
their needs. For the time being I offer an initial (incomplete)
implementation of a GalformRunner, plus an empty SharkRunner class as
well. The logic in GalformRunner comes from the original PSO C code, but
I might be getting some details wrong.

 * Most of the runtime options are not hardcoded into the code, but
rather passed down via command-line parameters. I say *most* and not
*all* because there are a few numbers that I hardcoded into the
GalformRunner class, which can be easily changed into runtime parameters
if required.

 * The main bits that need to be implemented from here on are: a) the
actual PSO, and b) the functions that calculate the .F term for each
particle.

I also have been able to do some limited manual testing on all this
code, mainly due to not having the proper setup here to run it
adequately. I trust this shouldn't be an issue when trying to running
the code in geryon or geryon2.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>